Swapped GPIO interface to use deprecated sysfs (#43)
sysfs is deprecated but I was unable to get GPIO working with /dev/mem. I would get no errors but it was not actually powering the pins. sysfs will be removed in 2020 at some point and has been deprecated for a number of years, but it works as a volume mount.